remove fetch commands and add chain and path fetching behind chains and paths cmd trees
